Assignment No. 3
Imagine a publishing company which does
marketing for book and audio cassette versions.
Create a class publication that stores the title (a
string) and price (type float) of publications.
From this class derive two classes: book which
adds a page count (type int) and tape which adds
a playing time in minutes (type float).Write a
program that instantiates the book and tape class,
allows user to enter data and displays the data
members. If an exception is caught, replace all the
data member values with zero values.