Assignment No. 7
Write a program in C++ to use map associative
container. The keys will be the names of states
and the values will be the populations of the
states. When the program runs, the user is
prompted to type the name of a state. The
program then looks in the map, using the state
name as an index and returns the population of
the state.